requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
cryptography
pytest
fastapi
uvicorn
sqlalchemy
cryptography
passlib
python-dotenv
pydantic
smtplib
pytest